requirements.txt
flask
mysql-connector-python
beautifulsoup4
curl-cffi
pyyaml